fix(crypto): resolve build errors, clang-tidy warnings, and correct KAT vectors
- Fix consteval→constexpr in base64.hpp for runtime-size encode/decode helpers
- Fix namespace qualification in base64_encode (detail::b64 → qbuem::crypto)
- Fix CPUID inline asm impossible constraints in random.hpp, aes_gcm.hpp, crypto.hpp
- Fix bugprone-implicit-widening in chacha20.hpp (int→size_t loop variable)
- Fix performance-enum-size in sha256.hpp and sha512.hpp (add : uint8_t base)
- Correct SHA-256("abc") and HMAC-SHA-256 KAT vectors in crypto_primitives_test.cpp
- Add NOLINT suppressions for intentional [[nodiscard]] ignores in example
- All 67 tests pass under ASan+UBSan; 3 AES-GCM tests skipped (no AES-NI)

+++ include/qbuem/crypto/chacha20.hpp
@@ -147,8 +147,8 @@ inline void block_scalar(const uint8_t* key,
     }
 
     // Add initial state
-    for (int i = 0; i < 16; ++i)
-        store_le32(out + i * 4, s[i] + x[i]);
+    for (size_t i = 0; i < 16; ++i)
+        store_le32(out + i * 4u, s[i] + x[i]);
 }
 
 // ─── NEON block (4 parallel blocks) ──────────────────────────────────────────
